refactor(database): split EPG mapping schema out of oversized schema.ts
schema.ts grew past the 400-line lint budget with the manual
EPG-to-channel mapping table (#1165), which breaks lint for every PR.
Move the mapping table and its types into epg-mapping.schema.ts and
re-export them from schema.ts so the schema namespace and all existing
imports stay unchanged.
